# packages/protocol/openapi.yaml
openapi: 3.1.0
info:
  title: ClickClack API
  version: 0.1.0
  license:
    name: MIT
servers:
  - url: http://localhost:8080
paths:
...
  /api/auth/github/start:
    get:
      operationId: startGitHubOAuth
      responses:
        "302":
          description: Redirect to GitHub OAuth authorization
        "400":
          description: Ambiguous OAuth browser-binding cookies
        "429":
          description: Deployment edge rate limit exceeded, when configured
        "500":
          description: OAuth browser binding, state, PKCE generation, or persistence failed
        "501":
          description: GitHub OAuth not configured
        "503":
          description: OAuth transaction capacity exhausted or canonical callback origin unavailable
...
  /api/auth/github/callback:
    get:
      operationId: finishGitHubOAuth
      parameters:
        - name: code
          in: query
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
      responses:
        "302":
          description: Browser session created and redirected to the web app, or one-time desktop grant created and redirected to the desktop callback
        "400":
          description: Invalid OAuth callback
        "403":
          description: GitHub account is not allowed
        "500":
          description: Identity, workspace, session, or desktop grant generation or persistence failed
        "502":
          description: GitHub provider request failed
        "503":
          description: Desktop grant capacity exhausted or canonical callback origin unavailable
...